server/requirements.txt
astroid==2.3.3
bcrypt==3.1.6
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.22.2
mccabe==0.6.1
pip-autoremove==0.9.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
requests==2.23.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
